InputForm 삭제
