--- src/components/MainPage/InputForm.module.scss
@@ -1,10 +1,0 @@
-.SubmitBtnContainer {
-  display: flex;
-  padding-top: 10px;
-  justify-content: flex-end;
-  font-family: "Jua-Regular";
-
-  .SubmitBtn {
-    margin-right: 10px;
-  }
-}
