--- src/components/RegisterPage/RegisterPage.scss
@@ -1,26 +1,0 @@
-@mixin btnStyle {
-  width: 80px;
-  height: 35px;
-  border-radius: 10px;
-  border: none;
-  background: #974dcb;
-  &:hover {
-    background: #6937a1;
-  }
-}
-
-button {
-  @include btnStyle();
-}
-
-.IdCheck {
-  display: flex;
-  margin-top: 10px;
-  justify-content: flex-end;
-  padding-bottom: 0px;
-}
-.RegiBtnContainer {
-  display: flex;
-  justify-content: flex-end;
-  margin-top: 10px;
-}
