+++ src/pages/RegisterPage.module.scss
@@ -1,0 +1,40 @@
+@mixin btnStyle {
+  width: 80px;
+  height: 35px;
+  border-radius: 10px;
+  border: none;
+  background: #974dcb;
+  &:hover {
+    background: #6937a1;
+  }
+}
+
+button {
+  @include btnStyle();
+}
+
+.RegisterPage {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 100px;
+  flex-direction: column;
+  font-family: "Jua-Regular";
+
+  .InputContainer {
+    margin-right: auto;
+    align-items: left;
+  }
+  .RegiBtnContainer {
+    display: flex;
+    justify-content: flex-end;
+    margin-top: 10px;
+  }
+}
+
+.IdCheck {
+  display: flex;
+  margin-top: 10px;
+  justify-content: flex-end;
+  padding-bottom: 0px;
+}
